Correção

+++ Trabalho 1/Trabalho 1/Image.cpp
@@ -158,41 +158,41 @@ Pixel Image::rgbToXYZ( Pixel rgb )
 {
     if (rgb[0] > 0.04045)
     {
-        rgb[0] = pow( ( rgb[0] + 0.055 ) / 1.055, 2.4 );
+        rgb[0] = (float)(pow( ( rgb[0] + 0.055 ) / 1.055, 2.4 ));
     }
     else
     {
-        rgb[0] = rgb[0] / 12.92;
+        rgb[0] = (float)(rgb[0] / 12.92);
     }
 
     if (rgb[1] > 0.04045)
     {
-        rgb[1] = pow( ( rgb[1] + 0.055 ) / 1.055, 2.4 );
+        rgb[1] = (float)(pow( ( rgb[1] + 0.055 ) / 1.055, 2.4 ));
     }
     else
     {
-        rgb[1] = rgb[1] / 12.92;
+        rgb[1] = (float)(rgb[1] / 12.92);
     }
 
     if (rgb[2] > 0.04045)
     {
-        rgb[2] = pow( ( rgb[2] + 0.055 ) / 1.055, 2.4 );
+        rgb[2] = (float)(pow( ( rgb[2] + 0.055 ) / 1.055, 2.4 ));
     }
     else
     {
-        rgb[2] = rgb[2] / 12.92;
+        rgb[2] = (float)(rgb[2] / 12.92);
     }
 
-    rgb[0] = rgb[0] * 100.0;
-    rgb[1] = rgb[1] * 100.0;
-    rgb[2] = rgb[2] * 100.0;
+    rgb[0] = (float)(rgb[0] * 100.0);
+    rgb[1] = (float)(rgb[1] * 100.0);
+    rgb[2] = (float)(rgb[2] * 100.0);
 
     Pixel XYZ;
 
     //Observer. = 2°, Illuminant = D65
-    XYZ[0] = rgb[0] * 0.4124 + rgb[1] * 0.3576 + rgb[2] * 0.1805;
-    XYZ[1] = rgb[0] * 0.2126 + rgb[1] * 0.7152 + rgb[2] * 0.0722;
-    XYZ[2] = rgb[0] * 0.0193 + rgb[1] * 0.1192 + rgb[2] * 0.9505;
+    XYZ[0] = (float)(rgb[0] * 0.4124 + rgb[1] * 0.3576 + rgb[2] * 0.1805);
+    XYZ[1] = (float)(rgb[0] * 0.2126 + rgb[1] * 0.7152 + rgb[2] * 0.0722);
+    XYZ[2] = (float)(rgb[0] * 0.0193 + rgb[1] * 0.1192 + rgb[2] * 0.9505);
 
     return XYZ;
 }
@@ -201,35 +201,35 @@ Pixel Image::rgbToXYZ( Pixel rgb )
 
 Pixel Image::XYZToLab( Pixel XYZ )
 {
-    XYZ[0] /= 95.047; //ref_X =  95.047   Observer= 2°, Illuminant= D65
-    XYZ[1] /= 100.000; //ref_Y = 100.000
-    XYZ[2] /= 108.883; //ref_Z = 108.883
+    XYZ[0] /= (float)95.047; //ref_X =  95.047   Observer= 2°, Illuminant= D65
+    XYZ[1] /= (float)100.000; //ref_Y = 100.000
+    XYZ[2] /= (float)108.883; //ref_Z = 108.883
 
     if (XYZ[0] > 0.008856)
     {
-        XYZ[0] = pow( XYZ[0], ( 1.0 / 3.0 ) );
+        XYZ[0] = (float)pow( XYZ[0], ( 1.0 / 3.0 ) );
     }
     else
     {
-        XYZ[0] = ( 7.787 * XYZ[0] ) + ( 16.0 / 116.0 );
+        XYZ[0] = (float)(( 7.787 * XYZ[0] ) + ( 16.0 / 116.0 ));
     }
 
     if (XYZ[1] > 0.008856)
     {
-        XYZ[1] = pow( XYZ[1], ( 1.0 / 3.0 ) );
+        XYZ[1] = (float)pow( XYZ[1], ( 1.0 / 3.0 ) );
     }
     else
     {
-        XYZ[1] = ( 7.787 * XYZ[1] ) + ( 16.0 / 116.0 );
+        XYZ[1] = (float)(( 7.787 * XYZ[1] ) + ( 16.0 / 116.0 ));
     }
 
     if (XYZ[2] > 0.008856)
     {
-        XYZ[2] = pow( XYZ[2], ( 1.0 / 3.0 ) );
+        XYZ[2] = (float)pow( XYZ[2], ( 1.0 / 3.0 ) );
     }
     else
     {
-        XYZ[2] = ( 7.787 * XYZ[2] ) + ( 16.0 / 116.0 );
+        XYZ[2] = (float)(( 7.787 * XYZ[2] ) + ( 16.0 / 116.0 ));
     }
 
     Pixel Lab;
@@ -250,35 +250,35 @@ Pixel Image::XYZTorgb( Pixel XYZ )
 
     Pixel rgb;
 
-    rgb[0] = XYZ[0] * 3.2406 + XYZ[1] * -1.5372 + XYZ[2] * -0.4986;
-    rgb[1] = XYZ[0] * -0.9689 + XYZ[1] * 1.8758 + XYZ[2] * 0.0415;
-    rgb[2] = XYZ[0] * 0.0557 + XYZ[1] * -0.2040 + XYZ[2] * 1.0570;
+    rgb[0] = (float)(XYZ[0] * 3.2406 + XYZ[1] * -1.5372 + XYZ[2] * -0.4986);
+    rgb[1] = (float)(XYZ[0] * -0.9689 + XYZ[1] * 1.8758 + XYZ[2] * 0.0415);
+    rgb[2] = (float)(XYZ[0] * 0.0557 + XYZ[1] * -0.2040 + XYZ[2] * 1.0570);
 
     if (rgb[0] > 0.0031308)
     {
-        rgb[0] = 1.055 * pow( rgb[0], ( 1 / 2.4 ) ) - 0.055;
+        rgb[0] = (float)(1.055 * pow( rgb[0], ( 1 / 2.4 ) ) - 0.055);
     }
     else
     {
-        rgb[0] = 12.92 * rgb[0];
+        rgb[0] = (float)(12.92 * rgb[0]);
     }
 
     if (rgb[1] > 0.0031308)
     {
-        rgb[1] = 1.055 * pow( rgb[1], ( 1 / 2.4 ) ) - 0.055;
+        rgb[1] = (float)(1.055 * pow( rgb[1], ( 1 / 2.4 ) ) - 0.055);
     }
     else
     {
-        rgb[1] = 12.92 * rgb[1];
+        rgb[1] = (float)(12.92 * rgb[1]);
     }
 
     if (rgb[2] > 0.0031308)
     {
-        rgb[2] = 1.055 * pow( rgb[2], ( 1 / 2.4 ) ) - 0.055;
+        rgb[2] = (float)(1.055 * pow( rgb[2], ( 1 / 2.4 ) ) - 0.055);
     }
     else
     {
-        rgb[2] = 12.92 * rgb[2];
+        rgb[2] = (float)(12.92 * rgb[2]);
     }
     return rgb;
 }
@@ -294,11 +294,11 @@ Pixel Image::LabToXYZ( Pixel Lab )
 
     if (pow( XYZ[1], 3 ) > 0.008856)
     {
-        XYZ[1] = pow( XYZ[1], 3 );
+        XYZ[1] = (float)pow( XYZ[1], 3 );
     }
     else
     {
-        XYZ[1] = ( XYZ[1] - 16.0 / 116.0 ) / 7.787;
+        XYZ[1] = (float)(( XYZ[1] - 16.0 / 116.0 ) / 7.787);
     }
 
     if (pow( XYZ[0], 3 ) > 0.008856)
@@ -307,7 +307,7 @@ Pixel Image::LabToXYZ( Pixel Lab )
     }
     else
     {
-        XYZ[0] = ( XYZ[0] - 16.0 / 116.0 ) / 7.787;
+        XYZ[0] = (float)(( XYZ[0] - 16.0 / 116.0 ) / 7.787);
     }
     if (pow( XYZ[2], 3 ) > 0.008856)
     {
@@ -315,12 +315,12 @@ Pixel Image::LabToXYZ( Pixel Lab )
     }
     else
     {
-        XYZ[2] = ( XYZ[2] - 16.0 / 116.0 ) / 7.787;
+        XYZ[2] = (float)(( XYZ[2] - 16.0 / 116.0 ) / 7.787);
     }
 
-    XYZ[0] = 95.047 * XYZ[0]; //ref_X =  95.047     Observer= 2°, Illuminant= D65
-    XYZ[1] = 100.000 * XYZ[1]; //ref_Y = 100.000
-    XYZ[2] = 108.883 * XYZ[2]; //ref_Z = 108.883
+    XYZ[0] = (float)(95.047 * XYZ[0]); //ref_X =  95.047     Observer= 2°, Illuminant= D65
+    XYZ[1] = (float)(100.000 * XYZ[1]); //ref_Y = 100.000
+    XYZ[2] = (float)(108.883 * XYZ[2]); //ref_Z = 108.883
 
     return XYZ;
 }
